+++ lib/models/money_objects/transactions/transactions.dart
@@ -4,6 +4,8 @@ import 'package:money/models/money_objects/accounts/account.dart';
 import 'package:money/models/money_objects/money_objects.dart';
 import 'package:money/models/money_objects/transactions/transaction.dart';
 
+export 'package:money/models/money_objects/transactions/transaction.dart';
+
 part 'transactions_csv.dart';
 
 part 'transactions_demo.dart';
@@ -43,5 +45,24 @@ class Transactions extends MoneyObjects<Transaction> {
       maxIdFound = max(maxIdFound, item.id.value);
     }
     return maxIdFound + 1;
+  }
+
+  Transaction? findExistingTransaction({
+    required final DateTime dateTime,
+    required final String payeeAsText,
+    required final String memo,
+    required final double amount,
+  }) {
+    // TODO make this more precises, at the moment we only match amount and date YYYY,MM,DD
+    return iterableList(true).firstWhereOrNull((transaction) {
+      if (transaction.amount.value == amount) {
+        if (transaction.dateTime.value?.year == dateTime.year &&
+            transaction.dateTime.value?.month == dateTime.month &&
+            transaction.dateTime.value?.day == dateTime.day) {
+          return true;
+        }
+      }
+      return false;
+    });
   }
 }